Alterar e criar arquivos para controller GET e POST transações

+++ DesafioBackEnd.API/Application/Command/Transacoes/TransacaoCommand.cs
@@ -5,8 +5,8 @@ namespace DesafioBackEnd.API.Application.Command.Transacoes
 {
     public class TransacaoCommand : IRequest<Transacao>
     {
-        public int? IdSender { get; set; }
-        public int? IdReceiver { get; set; }
+        public long? IdSender { get; set; }
+        public long? IdReceiver { get; set; }
         public decimal QuantiaTransferida { get; set; }
         public required DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     }
