+++ DesafioBackEnd.API/Application/Dto/Transacoes/CreateTransacaoDto.cs
@@ -5,10 +5,10 @@ namespace DesafioBackEnd.API.Application.Dto.Transacao
     public class CreateTransacaoDto
     {
         [Required(ErrorMessage = "O ID do pagador não pode ser vazio")]
-        public int? IdSender { get; set; }
+        public long? IdSender { get; set; }
 
         [Required(ErrorMessage = "O ID do recebedor não pode ser vazio")]
-        public int? IdReceiver { get; set; }
+        public long? IdReceiver { get; set; }
 
         [Required(ErrorMessage = "O valor a ser transferido não pode ser vazio")]
         public decimal QuantiaTransferida { get; set; }
